requirements.txt
python-dotenv
selenium
jpholiday
openai
# または
# transformers
# torch
